MacOS build: simplify maintenance of flags.sh (semgrep/semgrep-proprietary#2366)
The files `flags.sh` contain a list of libraries to be passed to the
linker. When adding a new tree-sitter language, this file had to be
edited otherwise, a cryptic error would be produced later in the build.
This PR automates the collection of libraries corresponding to
tree-sitter languages so that we don't have to know about `flags.sh`
when adding a new language.

synced from Pro 0c2daef3aa27d492d742bd6c2d9684e45ae347ba