chore: Allow begin/end in ocamlformat config (semgrep/semgrep-proprietary#2590)
The forced use of parentheses has been a consistent source of
frustration for me in this codebase. I find it particularly irritating
when there are a ton of close parentheses at the end of a line, making
it difficult to add additional code after that line and get it in the
right place.

For example:

```
match foo with
| A -> ...
| B -> (
     match bar with
       | C -> ...
       | D -> baz ())
```

This is a simple example, but we have significantly more complex ones
throughout our codebase. If I want to add something to the `D` case, I
have to make sure to put it before the close paren. If things are nested
more deeply, it makes it significantly more difficult to visually parse.

Fortunately, OCaml has a solution: `begin` and `end` can be used in
place of parentheses.

```
match foo with
| A -> ...
| B -> begin
     match bar with
       | C -> ...
       | D -> baz ()
  end
```

In cases like this, I find `begin/end` to be much more readable.

This change does not require any reformatting. It simply allows each
code author to use parentheses or begin/end, whichever makes sense to
them.

Test plan: `pre-commit run -a` -> no changes

synced from Pro 1bebe067294814a58f527d0edfc4580a12cfaab9