fix: disable telemetry in gc alarms (semgrep/semgrep-proprietary#2611)
In https://github.com/semgrep/semgrep-proprietary/pull/2492 we started
whole sale sending all logs via telemetry with levels > DEBUG when
tracing is enabled. There were some log statements in the memory limit,
inside the GC alarm. The opentelemetry library informs us that emitting
telemetry in a gc alarm can cause deadlocks!

To fix this, I've added a log tag that we can use to skip emitting
certain logs. This is also useful for if we want to log something and
just not emit it because it has sensitive info or similar. I double
checked and fixed the only gc alarm we create using this method.

## Test plan
```
semgrep scan --config p/default-v2 --error --trace --trace-endpoint semgrep-dev --pro --max-memory=3000
```
succeeds. Previously it would deadlock ~1/3 times on my machine

additionally I'm reintroducing
https://github.com/semgrep/semgrep-proprietary/pull/2531 in this PR, as
it is what caught this issue before, and I'm going to rerun the tests
that did a few times to ensure that it doesn't happen again. In these
tests the deadlock happened ~4/5 times, so if we get a few successful
runs in a row when this PR merges I will feel confident that this fix
works.

synced from Pro ceaaccdd1e065ad56a705783474e3af996749927