Disallow installing linux wheel on native windows (semgrep/semgrep-proprietary#2646)
Previously the linux wheel could be installed on windows but the
`semgrep` command would not be available on PATH since it was missing
the required `.exe` extension, rendering the installation
non-functional.

Since migrating to console_scripts to prepare for native windows
support, the `semgrep` command now has the required extension and is
being added to PATH. When run, this gives an error message that
`semgrep-core.exe` could not be found as the core binary does not have
the `.exe` extension in the linux wheel.

Address this by updating setup.py to correctly reflect that the linux
x86_64 wheel can only run on linux x86_64 platform/architecture. This
prevents package managers from considering this wheel for installation
on native windows.

synced from Pro 7a5f1c77dbdf2d90ea8d199ee549c160188b5936